% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Expit, i.e., the inverse of logit = log(x / (1-x))}
\usage{
expit(x)
}
\arguments{
\item{x}{numeric, possibly a vector}
}
\value{
expit(x)
}
\description{
Expit, i.e., the inverse of logit = log(x / (1-x))
}
\examples{
expit(-5)
}
