% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit of x for 0 <= x <= 1.}
\usage{
logit(p)
}
\arguments{
\item{p}{numeric between zero and one, possibly a vector}
}
\value{
logit of p
}
\description{
Logit of x for 0 <= x <= 1.
}
\examples{
logit(0.2)
}
