% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist}
\alias{pdist}
\title{Distance between points}
\usage{
pdist(x, coordsys = "euclidean")
}
\arguments{
\item{x}{x-y coordinates}

\item{coordsys}{either "euclidean" (default) or "wgs" for world geodetic
system.}
}
\value{
distance matrix
}
\description{
Compute distance matrix given a set of X-Y coordinates, possibly lat-long, using the \code{raster} package.
}
\examples{
m = matrix(1:6, 3, 2); pdist(m)
}
