\name{statsparty-package}
\alias{statsparty-package}
\alias{statsparty}
\docType{package}
\title{
\packageTitle{statsparty}
}
\description{
\packageDescription{statsparty}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{statsparty}
\packageIndices{statsparty}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{statsparty}

Maintainer: \packageMaintainer{statsparty}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
